+++ react/lib/JBImageInput.tsx
@@ -11,22 +11,24 @@ export type { JBImageInputConfig, JBImageInputBridge };
 
 export function JBImageInput<TValue>(props: Props<TValue>) {
   const element = useRef<JBImageInputWebComponent<TValue> | null>(null);
-  const { acceptTypes, ref, bridge, config, file, label, maxFileSize, message, multiple, name, required, validationList, value, onChange, onImageSelected, onInit, onLoad, onMaxSizeExceed, uploadType, ...otherProps } = props;
+  const { acceptTypes, ref, bridge, config, file, initialValue, label, maxFileSize, message, multiple, name, required, validationList, value, onChange, onImageSelected, onInit, onLoad, onMaxSizeExceed, uploadType, ...otherProps } = props;
   useImperativeHandle(ref,
     () => (element.current ?? undefined),
     [element]);
     
-  useJBImageInputAttribute(element, { acceptTypes, bridge, config, file, label, maxFileSize, message, multiple, name, required, validationList, value });
+  useJBImageInputAttribute(element, { acceptTypes, bridge, config, file, label, maxFileSize, message, multiple, name, required, validationList });
   useEvents(element, { onChange, onImageSelected, onInit, onLoad, onMaxSizeExceed });
 
   return (
-    <jb-image-input ref={element} upload-type={uploadType || 'AUTO'} {...otherProps}>
+    <jb-image-input ref={element} value={value ?? null} initialValue={initialValue ?? null} upload-type={uploadType || 'AUTO'} {...otherProps}>
       {props.children}
     </jb-image-input>
   );
 };
 type ImageInputProps<TValue> = EventProps<TValue> & JBImageInputAttributes<TValue> & {
   uploadType?: string,
+  value?: TValue | null,
+  initialValue?: TValue | null,
   ref?: ForwardedRef<JBImageInputWebComponent<TValue> | null | undefined>
 }
 export type Props<TValue> = ImageInputProps<TValue> & JBElementStandardProps<JBImageInputWebComponent, keyof ImageInputProps<TValue>>